refactor: auction page make public
