+++ dapp/src/auctions/components/AuctionItem.tsx
@@ -1,21 +1,17 @@
 // Copyright (c) 2025 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { Button, Card, CardType, LoadingIndicator } from '@iota/apps-ui-kit';
-import { useCurrentAccount, useIotaClient, useSignAndExecuteTransaction } from '@iota/dapp-kit';
+import { Card, CardType } from '@iota/apps-ui-kit';
 import { normalizeIotaName } from '@iota/iota-names-sdk';
-import { Transaction } from '@iota/iota-sdk/transactions';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { NameCard } from '@/components/name-card/NameCard';
 import { NameCardBody } from '@/components/name-card/NameCardBody';
 import { ExpiryDateIndicator } from '@/components/name-card/NameCardIndicators';
-import { queryKey } from '@/hooks/queryKey';
 
 import { getNameDisplaySrc } from '../../lib/utils/displayImage';
-import { useClaimAuctionTransaction } from '../hooks/useClaimAuctionTransaction';
 import { AuctionDetails } from '../hooks/useGetUserAuctions';
-import { getTimeRemaining, UserAuctionStatus } from '../lib/utils';
+import { UserAuctionStatus } from '../lib/utils';
+import { AuctionActionButton } from './AuctionActionButton';
 import { AuctionStatusBadge } from './AuctionStatusBadge';
 
 interface AuctionItemProps {
@@ -25,40 +21,9 @@ interface AuctionItemProps {
 }
 
 export function AuctionItem({ auction, auctionStatus, onBidClick }: AuctionItemProps) {
-    const iotaClient = useIotaClient();
-    const queryClient = useQueryClient();
-    const account = useCurrentAccount();
-
-    const { data: claimTransaction, isLoading: isClaimTransactionLoading } =
-        useClaimAuctionTransaction(account?.address || '', auction.name);
-
     const auctionDisplayImage = auction.metadata?.nftExpiration
         ? getNameDisplaySrc(auction.name, auction.metadata.nftExpiration.getTime())
         : null;
-
-    const { mutateAsync: signAndExecuteTransaction } = useSignAndExecuteTransaction();
-    const { mutateAsync: handleClaim, isPending: isSigningClaimTransaction } = useMutation({
-        async mutationFn(claimAuctionTransaction: Transaction) {
-            const transactionResult = await signAndExecuteTransaction({
-                transaction: claimAuctionTransaction,
-            });
-
-            await iotaClient.waitForTransaction({
-                digest: transactionResult.digest,
-            });
-        },
-        onSuccess() {
-            queryClient.invalidateQueries({
-                queryKey: queryKey.userAuctionHistory(account?.address),
-            });
-            // Invalidate auction so that it is detected as claimed
-            queryClient.invalidateQueries({ queryKey: queryKey.auctionMetadata(auction.name) });
-            // Refresh owned names so the claimed name appears
-            queryClient.invalidateQueries({
-                queryKey: queryKey.ownedObjects(account?.address || ''),
-            });
-        },
-    });
 
     if (!auction.metadata || auction.isLoading) {
         return (
@@ -83,44 +48,6 @@ export function AuctionItem({ auction, auctionStatus, onBidClick }: AuctionItemP
         );
     }
 
-    const timeRemaining = getTimeRemaining(auction.metadata);
-
-    const renderActionButton = () => {
-        if (auctionStatus === 'claimable') {
-            const isClaimLoading = isSigningClaimTransaction || isClaimTransactionLoading;
-
-            return (
-                <Button
-                    text={isSigningClaimTransaction ? 'Claiming...' : 'Claim'}
-                    icon={isClaimLoading ? <LoadingIndicator /> : null}
-                    onClick={() => {
-                        if (claimTransaction?.transaction) {
-                            handleClaim(claimTransaction.transaction);
-                        }
-                    }}
-                    disabled={isSigningClaimTransaction || !claimTransaction?.transaction}
-                    fullWidth
-                />
-            );
-        }
-
-        if (['outbid', 'top_bidder'].includes(auctionStatus) && timeRemaining > 0) {
-            return (
-                <Button
-                    text="Bid Again"
-                    onClick={() => {
-                        if (onBidClick) {
-                            onBidClick(auction.name);
-                        }
-                    }}
-                    fullWidth
-                />
-            );
-        }
-
-        return null;
-    };
-
     return (
         <NameCard name={auction.name} displaySrc={auctionDisplayImage}>
             <NameCardBody name={normalizeIotaName(auction.name)}>
@@ -129,7 +56,13 @@ export function AuctionItem({ auction, auctionStatus, onBidClick }: AuctionItemP
                     <AuctionStatusBadge status={auctionStatus} />
                 </div>
 
-                <div className="min-h-[44px] flex w-full">{renderActionButton()}</div>
+                <div className="min-h-[44px] flex w-full">
+                    <AuctionActionButton
+                        auction={auction}
+                        auctionStatus={auctionStatus}
+                        onBidClick={onBidClick}
+                    />
+                </div>
             </NameCardBody>
         </NameCard>
     );
