+++ dapp/src/auctions/components/AuctionPublicItem.tsx
@@ -1,13 +1,24 @@
+// Copyright (c) 2025 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
 import { Clock, IotaLogoSmall } from '@iota/apps-ui-icons';
-import { Button, ButtonType, Card, CardType, Divider, DividerType } from '@iota/apps-ui-kit';
+import { Card, CardType, Divider, DividerType } from '@iota/apps-ui-kit';
+import { useCurrentAccount } from '@iota/dapp-kit';
 import { normalizeIotaName } from '@iota/iota-names-sdk';
 
-import { AuctionDetails, formatTimeRemaining, getTimeRemaining } from '@/auctions';
+import {
+    AuctionDetails,
+    formatTimeRemaining,
+    getTimeRemaining,
+    getUserAuctionStatus,
+} from '@/auctions';
 import { useCountdown } from '@/auctions/hooks/useCountdown';
 import { NameCard } from '@/components/name-card/NameCard';
 import { NameCardBody } from '@/components/name-card/NameCardBody';
 import { formatNanosToIota } from '@/lib/utils';
 import { getNameDisplaySrc } from '@/lib/utils/displayImage';
+
+import { AuctionActionButton } from './AuctionActionButton';
 
 interface AuctionublicItemProps {
     auction: AuctionDetails;
@@ -18,6 +29,7 @@ export function AuctionPublicItem({ auction, onBidClick }: AuctionublicItemProps
     const auctionDisplayImage = auction.metadata?.nftExpiration
         ? getNameDisplaySrc(auction.name, auction.metadata.nftExpiration.getTime())
         : null;
+    const account = useCurrentAccount();
 
     if (!auction.metadata || auction.isLoading) {
         return (
@@ -31,59 +43,40 @@ export function AuctionPublicItem({ auction, onBidClick }: AuctionublicItemProps
         );
     }
 
+    const auctionStatus = getUserAuctionStatus(auction.metadata, account?.address || '');
+    const formattedPrice = formatNanosToIota(auction.metadata.currentBidNanos ?? BigInt(0));
     return (
         <NameCard name={auction.name} size="full" displaySrc={auctionDisplayImage}>
             <NameCardBody name={normalizeIotaName(auction.name)}>
-                <AuctionBidPrice
-                    price={auction.metadata?.currentBidNanos}
-                    auctionName={auction.name}
-                    onBidClick={onBidClick}
-                />
+                <div className="flex items-center justify-between">
+                    <div>
+                        <div className="text-label-md text-names-neutral-50">Current Bid</div>
+                        <div className="flex mt-1 items-center gap-2">
+                            <div className="bg-names-solid-blue rounded-full w-5 h-5 flex items-center justify-center">
+                                <IotaLogoSmall className="w-4 h-4" />
+                            </div>
+                            <div className="flex items-baseline gap-1">
+                                <span className="text-body-lg">{formattedPrice}</span>
+                                {/* <span className="text-label-md text-names-neutral-70">
+                            ${convertedToFiat.toString()}
+                        </span> */}
+                            </div>
+                        </div>
+                    </div>
+                    <div>
+                        <AuctionActionButton
+                            auction={auction}
+                            auctionStatus={auctionStatus}
+                            onBidClick={onBidClick}
+                        />
+                    </div>
+                </div>
                 <div className="my-4">
                     <Divider type={DividerType.Horizontal} />
                 </div>
                 <AuctionTimeRemaining auction={auction} />
             </NameCardBody>
         </NameCard>
-    );
-}
-
-function AuctionBidPrice({
-    price,
-    auctionName,
-    onBidClick,
-}: {
-    price?: bigint;
-    auctionName: string;
-    onBidClick: (name: string) => void;
-}): JSX.Element {
-    const formattedPrice = formatNanosToIota(price ?? BigInt(0));
-
-    return (
-        <div className="flex items-center justify-between">
-            <div>
-                <div className="text-label-md text-names-neutral-50">Current Bid</div>
-                <div className="flex mt-1 items-center gap-2">
-                    <div className="bg-names-solid-blue rounded-full w-5 h-5 flex items-center justify-center">
-                        <IotaLogoSmall className="w-4 h-4" />
-                    </div>
-                    <div className="flex items-baseline gap-1">
-                        <span className="text-body-lg">{formattedPrice}</span>
-                        {/* <span className="text-label-md text-names-neutral-70">
-                            ${convertedToFiat.toString()}
-                        </span> */}
-                    </div>
-                </div>
-            </div>
-            <div>
-                <Button
-                    text="Bid"
-                    onClick={() => onBidClick(auctionName)}
-                    fullWidth
-                    type={ButtonType.Outlined}
-                />
-            </div>
-        </div>
     );
 }
